chore: rename internal HTML plugin to RsbuildHtmlPlugin
